requirements.txt
absl-py==0.8.1
albumentations==0.4.4
astor==0.8.0
beautifulsoup4==4.8.1
boto==2.49.0
boto3==1.10.30
botocore==1.13.30
bs4==0.0.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
efficientnet==1.0.0
gast==0.2.2
gensim==3.8.1
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
image-classifiers==1.0.0
imageio==2.6.1
imgaug==0.2.6
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.0
PyWavelets==1.1.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.4.1
segmentation-models==1.0.0
six==1.13.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
tensorboard==2.1.1
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
tta-wrapper==0.0.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
